+++ src/main/scala/ncbiTaxonomy/scientificName.scala
@@ -1,0 +1,8 @@
+package com.bio4j.data.ncbiTaxonomy
+
+// Only rows with scientific names:
+case class ScientificName(val scientificLine: String) extends AnyVal {
+
+  def taxID: String = scientificLine.dmpColumns(0)
+  def scientificName: String = scientificLine.dmpColumns(1)
+}
